requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0

